+++ jupiter/src/renderer/opengl/GLRenderer.h
@@ -14,7 +14,7 @@ public:
 
     void clear() const override;
     void setClearColor(const jm::Vec4f& color) const override;
-    void setViewport(const float32 width, const float32 height) const override;
+    void setViewport(const uint32 width, const uint32 height) const override;
 
     void drawElements(const SharedPtr<VertexArray>& vertArray) const override;
 };
